Drinks page
This page has drink options using html, css, and javascript. I am using the <select> element with options and taking the value selected to use javascript manipulation. there is also an addCart icon that appears above that toggles colors when added to cart. The images src wil change and a different picture comes up based on the value selected.